+++ gitlab/v4/objects/branches.py
@@ -1,6 +1,4 @@
-from typing import Any, cast, Union
-
-from gitlab.base import RESTManager, RESTObject
+from gitlab.base import RESTObject
 from gitlab.mixins import (
     CRUDMixin,
     NoUpdateMixin,
@@ -22,23 +20,18 @@ class ProjectBranch(ObjectDeleteMixin, RESTObject):
     _id_attr = "name"
 
 
-class ProjectBranchManager(NoUpdateMixin, RESTManager):
+class ProjectBranchManager(NoUpdateMixin[ProjectBranch]):
     _path = "/projects/{project_id}/repository/branches"
     _obj_cls = ProjectBranch
     _from_parent_attrs = {"project_id": "id"}
     _create_attrs = RequiredOptional(required=("branch", "ref"))
-
-    def get(
-        self, id: Union[str, int], lazy: bool = False, **kwargs: Any
-    ) -> ProjectBranch:
-        return cast(ProjectBranch, super().get(id=id, lazy=lazy, **kwargs))
 
 
 class ProjectProtectedBranch(SaveMixin, ObjectDeleteMixin, RESTObject):
     _id_attr = "name"
 
 
-class ProjectProtectedBranchManager(CRUDMixin, RESTManager):
+class ProjectProtectedBranchManager(CRUDMixin[ProjectProtectedBranch]):
     _path = "/projects/{project_id}/protected_branches"
     _obj_cls = ProjectProtectedBranch
     _from_parent_attrs = {"project_id": "id"}
@@ -57,7 +50,26 @@ class ProjectProtectedBranchManager(CRUDMixin, RESTManager):
     )
     _update_method = UpdateMethod.PATCH
 
-    def get(
-        self, id: Union[str, int], lazy: bool = False, **kwargs: Any
-    ) -> ProjectProtectedBranch:
-        return cast(ProjectProtectedBranch, super().get(id=id, lazy=lazy, **kwargs))
+
+class GroupProtectedBranch(SaveMixin, ObjectDeleteMixin, RESTObject):
+    _id_attr = "name"
+
+
+class GroupProtectedBranchManager(CRUDMixin[GroupProtectedBranch]):
+    _path = "/groups/{group_id}/protected_branches"
+    _obj_cls = GroupProtectedBranch
+    _from_parent_attrs = {"group_id": "id"}
+    _create_attrs = RequiredOptional(
+        required=("name",),
+        optional=(
+            "push_access_level",
+            "merge_access_level",
+            "unprotect_access_level",
+            "allow_force_push",
+            "allowed_to_push",
+            "allowed_to_merge",
+            "allowed_to_unprotect",
+            "code_owner_approval_required",
+        ),
+    )
+    _update_method = UpdateMethod.PATCH