+++ tests/functional/api/test_registry.py
@@ -11,10 +11,10 @@ def protected_registry_feature(gl: Gitlab):
 
 @pytest.mark.skip(reason="Not released yet")
 def test_project_protected_registry(project: Project):
-    rules = project.registry_protection_rules.list()
+    rules = project.registry_protection_repository_rules.list()
     assert isinstance(rules, list)
 
-    protected_registry = project.registry_protection_rules.create(
+    protected_registry = project.registry_protection_repository_rules.create(
         {
             "repository_path_pattern": "test/image",
             "minimum_access_level_for_push": "maintainer",
@@ -26,3 +26,8 @@ def test_project_protected_registry(project: Project):
     protected_registry.minimum_access_level_for_push = "owner"
     protected_registry.save()
     assert protected_registry.minimum_access_level_for_push == "owner"
+
+    protected_registry.delete()
+
+    rules = project.registry_protection_repository_rules.list()
+    assert rules == []